src/pyproject.toml
[project]
name = "freebie"
version = "0.1.0"
description = ""
authors = [
    {name = "EvaReboucas30",email = "140523579+EvaReboucas30@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "freebie", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
